Move interfaces to the test folder
